% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{cc}
\alias{\%,\%}
\alias{cc}
\title{Comma Concatenation}
\usage{
cc(...)

x \%,\% y
}
\arguments{
\item{...}{Character vectors.}

\item{x}{Character vector.}

\item{y}{Character vector.}
}
\value{
A length-one character vector in which each element in `...` is
separated by a comma (and a space).
}
\description{
`cc` collapses text into a comma-separated list (the colloquial kind of
list).
}
\examples{
cc(LETTERS[1:11])
# [1] "A, B, C, D, E, F, G, H, I, J, K"

}
\seealso{
`\link{cc_or}` `\link{cc_and}`
}

