% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{cc_and}
\alias{\%and\%}
\alias{cc_and}
\title{Comma Concatenation with a Penultimate "and"}
\usage{
cc_and(..., oxford = FALSE)

x \%and\% y
}
\arguments{
\item{...}{Character vectors.}

\item{oxford}{Whether to use the Oxford comma.}

\item{x}{Character vector.}

\item{y}{Character vector.}
}
\description{
`cc_and` collapses text into a comma-separated list inserting`"and"` before
the last element of text.
}
\seealso{
`\link{cc}` `\link{cc_or}`
}

