% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{cc_or}
\alias{\%or\%}
\alias{cc_or}
\title{Comma Concatenation with a Penultimate "or"}
\usage{
cc_or(..., oxford = FALSE)

x \%or\% y
}
\arguments{
\item{...}{Character vectors.}

\item{oxford}{Whether to use the Oxford comma.}

\item{x}{Character vector.}

\item{y}{Character vector.}
}
\description{
`cc_or` collapses text into a comma-separated list inserting`"or"` before the
last element of text.
}
\examples{
a <- "one thing"
b <- "another"
message("likely reason: ", cc_or(a, b))
}
\seealso{
`\link{cc}` `\link{cc_and}`
}

