% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{grapes-plus-grapes}
\alias{\% + \%}
\alias{\%+\%}
\alias{\%,\%}
\alias{\%and\%}
\alias{\%or\%}
\alias{grapes-plus-grapes}
\title{Binary Concatenation}
\usage{
x \%+\% y

x \% + \% y

x \%,\% y

x \%or\% y

x \%and\% y
}
\arguments{
\item{x, y}{Character vectors.}
}
\description{
\code{\%+\%} is a binary \code{paste} like Python's \code{+} for strings.
\code{\% + \%}: inserts a space between its inputs.  \code{\%,\%},
\code{\%or\%} and \code{\%and\%} are binary versions of \code{\link{cc}},
\code{\link{cc_or}}, and \code{\link{cc_and}}.
}
\examples{
v <- "important value"
v \%+\% "!"

message("Two" \% + \% "words")
}

