% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.r
\name{infix}
\alias{\% c \%}
\alias{\%,\%}
\alias{\%and\%}
\alias{\%c\%}
\alias{\%or\%}
\alias{infix}
\title{Binary Infix Concatenation}
\usage{
x \%c\% y

x \% c \% y

x \%,\% y

x \%or\% y

x \%and\% y
}
\arguments{
\item{x, y}{Character vectors.}
}
\description{
\describe{
\item{\code{\%c\%}:}{binary infix operator for strings.}
\item{\code{\% c \%}:}{like \code{\%c\%} but with a space between its
inputs.}
\item{\code{\%,\%}, \code{\%or\%}, \code{\%and\%}:}{infix versions of
      \code{\link{cc}}, \code{\link{cc_or}}, \code{\link{cc_and}}.}
}
}
\examples{
v <- "important value"
v \%c\% "!"

message("Two" \% c \% "words")
}

