% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{\%+\%}
\alias{\% + \%}
\alias{\%+\%}
\title{Binary Concatenation}
\usage{
x \%+\% y

x \% + \% y
}
\arguments{
\item{x, y}{Character vector.}
}
\description{
`%+%` is a binary `paste` like Python's `+` for strings.

`% + %` is like `%+%` but inserts a space between its inputs.
}
\examples{
v <- "important value"
v \%+\% "!"
message("Two" \% + \% "words")
}

