% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.r
\name{infix}
\alias{\% + \%}
\alias{\%+\%}
\alias{\%,\%}
\alias{\%and\%}
\alias{\%or\%}
\alias{infix}
\title{Binary Infix Concatenation}
\usage{
x \%+\% y

x \% + \% y

x \%,\% y

x \%or\% y

x \%and\% y
}
\arguments{
\item{x, y}{Character vectors.}
}
\description{
\describe{
\item{\code{\%+\%}:}{binary infix operator for strings.}
\item{\code{\% + \%}:}{like \code{\%+\%} but with a space between its
inputs.}
\item{\code{\%,\%}, \code{\%or\%}, \code{\%and\%}:}{infix versions of
      \code{\link{cc}}, \code{\link{cc_or}}, \code{\link{cc_and}}.}
}
}
\examples{
v <- "important value"
v \%+\% "!"

message("Two" \% + \% "words")
}

